Add files via upload
Overview: In pursuit of contributing to automotive safety, I undertook a comprehensive analysis of NHSTA data on car crashes, specifically focusing on substance involvement. The resulting project is a dynamic Tableau dashboard designed to provide valuable insights into the impact of substances on car safety.

Key Features:

Substance Metrics Comparison:

Parameter-driven switch between alcohol and drug metrics, enabling a detailed comparative analysis.

Weekly Crash Frequency:

Visual representation of crash frequency throughout the week, offering insights into substance-related patterns.

Injury Impact Visualization:

Graphic display of the impact of substance-related crashes on the severity of injuries sustained.

Crash Characteristics Analysis:

Examination of crash characteristics and their correlation with hospitalization or fatality, categorized by alcohol or drug involvement.

This project serves as a testament to my commitment to leveraging data for informed decision-making in the realm of automotive safety. It is not only a showcase of technical skills but a demonstration of strategic thinking and analytical prowess.